[RO-3244] Link a role requirement file for rpc-differ
The rpc-differ application expects to find the
"ansible-role-requiremnts.yml" file in the root path of the repo. This
change make diff generation tooling link a product requirements file to
expected path when executed. Because this change has the potential to
create a modified git tree, the "ansible-role-requirements.yml" has been
added to .gitignore.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>